dnsdist: Handle HTTP/3 error responses
